+++ CMakeLists.txt
@@ -1,9 +1,8 @@
 cmake_minimum_required(VERSION 3.13)
 
 project(Another-Card-Game-ACG)
-set(CMAKE_CXX_STANDARD 17)
-
-set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} ")
+set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -Wall")
 
 if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
     # using GNU
@@ -11,9 +10,6 @@ if ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
 endif()
 
 include_directories(lib)
-include_directories(src)
-include_directories(src/pages)
-
 
 add_subdirectory(src)
 option (DO_TEST "Running test" N)
@@ -28,10 +24,5 @@ if (DO_TEST)
 #    target_link_libraries(test_main gtest gtest_main)
 endif(DO_TEST)
 
-
-
-#file(COPY data DESTINATION .)
-
-add_executable(main main.cpp test/FunctionalTest.cpp)
+add_executable(main main.cpp)
 target_link_libraries(main ACGMain)
-
