+++ src/ACGMain.h
@@ -5,27 +5,22 @@
 #ifndef ANOTHER_CARD_GAME_ACG_ACGMAIN_H
 #define ANOTHER_CARD_GAME_ACG_ACGMAIN_H
 
-
-#include "AccountSystemController.h"
-
+#include "AccountSystem.h"
 #include "Reader.h"
 #include "Printer.h"
 #include "Router.h"
 
 class ACGMain {
 private:
-
-    AccountSystemController *accountSystemController;
+    AccountSystem *accountSystem;
     Reader * reader;
     Printer * printer;
     Router * router;
-    void read();
-    void print();
-    void route();
 public:
     ACGMain();
     void run();
-
+    void read();
+    void print();
 };
