--- src/AccountSystemController.cpp
@@ -1,5 +1,0 @@
-//
-// Created by jacky on 2019-05-11.
-//
-
-#include "AccountSystemController.h"
