+++ src/CMakeLists.txt
@@ -1,7 +1,1 @@
-
-add_library(ACGMain ACGMain.cpp Printer.cpp
-        Reader.cpp AccountSystem.cpp Router.cpp
-        LadderSystem.cpp AccountSystemParser.cpp
-        AccountSystemController.cpp
-        Room.cpp Deck.cpp desk.cpp)
-
+add_library(ACGMain ACGMain.cpp Printer.cpp Reader.cpp AccountSystem.cpp Router.cpp LadderSystem.cpp)
