+++ src/Printer.h
@@ -9,11 +9,9 @@
 #include <queue>
 #include <mutex>
 
-
 #include "nlohmann/json.hpp"
 
 class Printer {
-
 public:
     Printer(std::queue<nlohmann::json> &toPrintQueue, std::mutex &p);
 
@@ -21,10 +19,8 @@ public:
 
 private:
     std::queue<nlohmann::json> *toPrintQueue;
-
     std::mutex * queueMutex;
 };
-
