+++ src/Reader.h
@@ -6,21 +6,18 @@
 #define ANOTHER_CARD_GAME_ACG_READER_H
 
 #include <string>
-
-#include <deque>
+#include <queue>
 #include <mutex>
 
-#include <nlohmann/json.hpp>
-using json = nlohmann::json;
-class Reader
-{
+#include "nlohmann/json.hpp"
+
+class Reader {
 public:
-    Reader(std::deque<nlohmann::json> &toDeliver, std::mutex &mutex);
+    Reader(std::queue<nlohmann::json> &toDeliver, std::mutex &mutex);
     void read();
-    json popJson(std::string type, int id);
 
 private:
-    std::deque<nlohmann::json> *toDeliverQueue;
+    std::queue<nlohmann::json> *toDeliverQueue;
     std::mutex *queueMutex;
 };
