--- src/page/DeckPage.h
@@ -1,14 +1,0 @@
-//
-// Created by jacky on 2019/4/20.
-//
-
-#ifndef ANOTHER_CARD_GAME_ACG_DECKPAGE_H
-#define ANOTHER_CARD_GAME_ACG_DECKPAGE_H
-
-
-class DeckPage {
-
-};
-
-
-#endif //ANOTHER_CARD_GAME_ACG_DECKPAGE_H
