--- src/page/PageBase.h
@@ -1,16 +1,0 @@
-//
-// Created by jacky on 2019/4/20.
-//
-
-#ifndef ANOTHER_CARD_GAME_ACG_PAGEBASE_H
-#define ANOTHER_CARD_GAME_ACG_PAGEBASE_H
-
-
-class PageBase {
-public:
-    virtual void updatePage();
-    virtual string getName();
-};
-
-
-#endif //ANOTHER_CARD_GAME_ACG_PAGEBASE_H
