--- src/page/PageController.cpp
@@ -1,5 +1,0 @@
-//
-// Created by jacky on 2019/4/20.
-//
-
-#include "page/PageController.h"
