--- test/FunctionalTest.cpp
@@ -1,4 +1,0 @@
-//
-// Created by jacky on 2019-05-11.
-//
-
