--- testCase/login.json
@@ -1,13 +1,0 @@
-{
-  "time": 1,
-  "type": 123,
-  "userId": 123,
-  "deviceId":  123,
-  "data" : {
-    "eventType": "asd",
-    "x" : [0,1,2,3,4],
-    "s" : ["123", "245", "456"],
-    "params" : [948794, [1,23,4,5,6], 66, [123,4,56,7]]
-  },
-  "//d" : "//“params”: [94879487, [1,2,3,...], 77, [123,513,23,....]]"
-}
